Finished MoPub integration Added decimal support fro Desartes Upgraded to sdk24 Begun change to vectors

+++ app/src/main/java/com/example/danny/geometryhelper/DescartesSigns.java
@@ -7,7 +7,7 @@ public class DescartesSigns {
 
     private final AlgebraParse parser;
 
-    private int[] coeffs;
+    private double[] coeffs;
 
     private int[] simpleCoeffs;
